Resolving comments
